src/net/endpoint.ts => src/net/endpoint/{index,factory,helpers}.ts
