+++ src/cli/command/start/config-file-parser.ts
@@ -4,7 +4,7 @@ import * as path from 'path';
 import { ICluster } from '../../../cluster';
 import * as logger from '../../../logger';
 import { ICreateServerOptions } from '../../../server';
-import { IEndpoint, createEndpoint } from '../../../net';
+import { IEndpoint, createEndpoint } from '../../../net/endpoint';
 
 // A config file parser that accepts a file path,
 // validates the contents, and returns a configuration
