+++ src/cluster.ts
@@ -1,4 +1,4 @@
-import { IEndpoint } from './net';
+import { IEndpoint } from './net/endpoint';
 
 export interface ICluster {
     readonly servers: { [id: string]: IEndpoint };
