+++ test/server/rpc/rpc-service.test.ts
@@ -2,7 +2,7 @@ import { expect } from 'chai';
 
 import { IRpcService } from '../rpc';
 import * as AppendEntries from '../rpc/message/append-entries';
-import { IEndpoint, createEndpoint } from '../../net';
+import { IEndpoint, createEndpoint } from '../../net/endpoint';
 import { createRpcService } from '../rpc';
 
 describe('rpc service', function() {
