+++ test/server/state/candidate.test.ts
@@ -7,7 +7,7 @@ import sinon from 'sinon';
 
 import * as AppendEntries from '../rpc/message/append-entries';
 import * as RequestVote from '../rpc/message/request-vote';
-import { IEndpoint, createEndpoint } from '../../net';
+import { IEndpoint, createEndpoint } from '../../net/endpoint';
 import { IRpcEventListener, IRpcService, createRpcService } from '../rpc';
 import { IServer, ServerId, createServer } from '../';
 import { ITimer, createTimer, createTimeoutChooser } from '../timer';
